(1) changed three column names in tasks table. (2) added new routes to routes.eb (show_admin, user_complete, admin_complete). (3) added a new view in events directory to handle admin functionality. It might not be as DRY, but it improves readability. (4) made _table_row & _table_headers partials to DRY the show views. (5) added methods to task.rb to handle point calculation and distribution. (6) added helper methods for tasks that handle checkbox genration, the join/leave link for a user, and to check if a user is an admin. (7) added actions user_complete, admin_complete to handle the selection of multiple checkboxes by a user or an admin. (8) changed the names of two permitted task params. (9) added a get show_admin action to the events controller.

+++ app/helpers/tasks_helper.rb
@@ -7,22 +7,44 @@ module TasksHelper
   # then list out of the participants and provide a leave task with a redirect to the user_task destroy action
   # otherwise (if it does not contain the pa) just list out the participants
 
+  def user_is_event_admin(event)
+    event.admin_id == current_user.id
+  end
+
   def check_participant_join_status(task)
-    if task.users.where("username = ?", current_user.username).first
-      if task.users.size <= task.max_participants
-        link_to "Leave Task", users_task_path(task), method: :delete
-      end
+    if !task.completed
+      if task.users.where("username = ?", current_user.username).first
+        if task.users.size <= task.max_participants
+          link_to "Leave Task", users_task_path(task), method: :delete
+        end
 
-    elsif task.users.size < task.max_participants
-      if !task.users.where("username = ?", current_user.username).first
-        link_to "Join Task", users_tasks_path(task, id: task.id), method: :post
-      end
+      elsif task.users.size < task.max_participants
+        if !task.users.where("username = ?", current_user.username).first
+          link_to "Join Task", users_tasks_path(task, id: task.id), method: :post
+        end
 
-    else
-      "Task Full"
+      else
+        "Task Full"
+      end
     end
   end
 
+
+  def display_user_checkbox(task)
+    if task.users.where("username = ?", current_user.username).first
+      if task.user_completed_at && task.admin_confirmed_completion_at
+        "Complete"
+      elsif task.user_completed_at && !task.admin_confirmed_completion_at
+        "Pending Confirmation"
+      else
+        check_box_tag "user_task_ids[]", task.id
+      end
+    end
+  end
+
+  def display_admin_checkbox(task)
+    check_box_tag "admin_task_ids[]", task.id
+  end
 
 
 end
